+++ extensions/twitter/src/send.tsx
@@ -1,6 +1,9 @@
+import { withXAuth } from "./v2/lib/with_x_auth";
 import { ReactElement } from "react";
 import { TweetSendThreadFormV2 } from "./v2/components/send";
 
-export default function SendTweetRoot({ launchContext }: { launchContext?: { defaultValue: string } }): ReactElement {
+function SendTweetRoot({ launchContext }: { launchContext?: { defaultValue: string } }): ReactElement {
   return <TweetSendThreadFormV2 defaultValue={launchContext?.defaultValue} />;
 }
+
+export default withXAuth(SendTweetRoot);